adapt changes for sqlite test

+++ FIK.DAL/FIK.ORM.Tests.Share/Models/Sales.cs
@@ -6,8 +6,8 @@ namespace FIK.ORM.Tests.Share.Models
 {
     public class Sales
     {
-        public int Id { get; set; }
-        public int ItemId { get; set; }
+        public long Id { get; set; }
+        public long ItemId { get; set; }
         public decimal Quantity { get; set; }
         public decimal TotalPrice { get; set; }
     }
